requirements.txt
beautifulsoup4==4.8.0
python-dotenv==0.10.3
selenium==3.141.0
